requirements.txt

Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
fastparquet==0.7.1
Flask==2.0.1
GDAL=3.3.1
geos==0.2.3
numpy==1.21.2
pandas==1.3.2
pipreqs==0.4.10
postgis==1.0.4
psycopg2==2.9.1
pyarrow==5.0.0
